Implement proper CRUD for event administration
The former continuous form like event administration certainly has its
advantages for a relatively small amount of events, but does not scale
well.  While the CRUD alternative implemented so far does not scale
much better, it is at least improvable (sort, search, pagination).

Furthermore, the old variant makes it hard to report invalid input;
while we do not do that yet for the CRUD variant, it is obviously
simpler to do.

Finally, the offline lock implemented for the old variant leaves much
to be desired, i.e. it is not fine grained enough.  This is resolved by
switching to proper CRUD.