Close #41: Drop JS enhancements for pre ES6 browsers
For now we're happy to "cut the mustard" by using ES modules, and to
remove the superfluous checking for HTML5 browsers, as well as running
the code on DOMContentLoaded, since ES modules are deferred anyway.

We also put the scripts into the `<head>` instead of the end of the
`<body>`; `$bjs` was good when it had been introduced, but no longer
is apt.