Fix #58: Calendar module not loaded if only called in the template
We should load JS modules as early as possible, but since `$hjs` may
have been output already, we need to load the module from the template,
too.

Since we don't have the need for "proper" modules (i.e. distinguish
between declarations and statements) for now, we move the widget
initialization to the respective modules.  This avoids the need for
once-routines in PHP, because a JS module is only evaluated once at
most.