Don't pass plugincall arguments into controller constructors
These are not dependencies of the controller, but should rather be
arguments to the action(s).

While that requires us to pass a couple of arguments around (and we
even uses some reference parameters for simplicity), this is actually
a good thing, since it shows us that there is ample room for cleaning
up.