Close #62: Improve display of event editor overview
While a lot can be improved by customizing the stylesheet, we should
provide a more pleasant experience by default.

The overview should be dense, so users can easily see even a somewhat
large amounts of events (filtering would be necessary for very many
events, though).  Thus, we drop the end-date column altogether, but add
the start-time to the start-date column.  We also replace the text
buttons with icons (still in an accessible way).  Finally we switch to
fixed table layout, and cut eventual text-overflow in the summary
column.  That should be sufficient to reduce the height of the rows to
a single line (may still need some CSS tweaks for some templates).

The edit form of a single event doesn't profit from flex layout in any
way.  There is not much info, so we can stick with a more classic
layout which looks way more pleasant.  We also drop the time labels
(we likely want to switch to datetime-local inputs anyway).