Introduce occurrence concept
An occurrence is an `Event` which is a single occurrence of a recurring
event.  Since we only support birthdays for now (a special cased yearly
recurring event), we can keep that super simple.  We change the three
scheduling methods of `Event` to return an occurrence (or `null`),
which is a modified clone of birthday events, but the actual event
otherwise.  We also rename these methods to reflect the behavioral
change.