Support for editing a single event occurrence
Recurring events may have exceptions, e.g. a single occurrence of a
weekly event may be moved because it falls on a holiday.  To avoid too
complex logic modeling this, we support splitting an recurring event in
up to three parts: the exceptional event, and possibly the recurring
event up to this date, and the recurring event starting with the next
occurrence.  Afterwards, all new events can be edited like before, and
split again, if necessary.